Engine: Moved unit tests to their own directory
