+++ engine/tests/tests.cpp
@@ -1,2 +1,2 @@
-#include "rotation.cpp"
-#include "vector_utils.cpp"
+#include "unit_tests/rotation.cpp"
+#include "unit_tests/vector_utils.cpp"
